remove subclasses of site
